move smile config decoding and initialisation into root view model. add a closure to send back update config to root view.
